src/requirements.txt
appdirs==1.4.0
cycler==0.10.0
elasticsearch==1.9.0
elasticsearch-dsl==0.0.11
future==0.16.0
httpie==0.9.9
matplotlib==2.0.0
numpy==1.12.1
oauthlib==2.0.1
packaging==16.8
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-twitter==3.2.1
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
urllib3==1.20
